Added a template trap handler
=============================
A new section called trap_handler has been added, which is jumped to
after the registers have been saved to the stack and the trap_handler
jumps to trap_return after completing handling.

Adding this section makes it easier for a user to edit the trap_handling
code to their liking.